+++ javersion-jdbc/src/test/resources/db/migration/common/V1__common_base.sql
@@ -1,11 +1,1 @@
-create table VERSION_TYPE (
-  NAME varchar(8),
-  primary key (NAME)
-);
-insert into VERSION_TYPE values ('NORMAL');
-insert into VERSION_TYPE values ('RESET');
-
-create table REPOSITORY (
-  ID varchar(32) not null,
-  primary key (ID)
-);
+-- Shared VERSION_TYPE and REPOSITORY tables are no longer needed